icp-qual/ad-intel/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
